+++ src/vinifera/vinifera_defines.h
@@ -262,6 +262,7 @@ typedef enum ExtTEventType
     EXT_TEVENT_LOCAL_LESS_THAN_GLOBAL,
     EXT_TEVENT_LOCAL_LESS_THAN_LOCAL,
     EXT_TEVENT_BUILDING_DOES_NOT_EXIST,
+    EXT_TEVENT_DESTROYED_ONLY,
 
     /**
      *  The new total ExtTEventType count.